Hopefully this works (Stuff for monday)
The modular encoder distance is negative when moving forward. This causes the error which I calculate in the variable currentDistanceAwayFromTarget to increase, which not only causes the PID (Proportional) controller to increase the speed, but also causes the robot to not stop at its intended target. This perfectly explains what I observed yesterday.